+++ R/acled.api.R
@@ -23,7 +23,7 @@
 #' @param start.date character string. Supply the earliest date to be retrieved. Format: "yyyy-mm-dd".
 #' @param end.date character string. Supply the last date to be retrieved. Format: "yyyy-mm-dd".
 #' @param add.variables character vector. Supply the names of ACLED variables you wish to add to the
-#' default output (see [ACLED's codebook](https://acleddata.com/resources/general-guides/) for details). The default
+#' default output (see [ACLED's codebook](https://acleddata.com/knowledge-base/codebook/) for details). The default
 #' output includes: event_id_cnty, region, country, year, event_date, source, admin1, admin2, admin3, location, latitude, longitude, event_type, sub_event_type,
 #' interaction, fatalities, tags, and the download timestamp.
 #' @param all.variables logical. When set to FALSE (default), a narrow default selection of variables is returned (which
@@ -32,7 +32,8 @@
 #' @param dyadic logical. When set to FALSE (default), monadic data is returned (one
 #' observation per event). If set to TRUE, dyadic data is returned.
 #' @param interaction numeric vector. Supply one or more interaction codes to narrow down which events should be
-#' retrieved (see [ACLED's codebook](https://acleddata.com/resources/general-guides/) for details.
+#' retrieved. See [ACLED's codebook](https://acleddata.com/knowledge-base/codebook/) for information on which
+#' interactions are represented by which numeric codes.
 #' @param other.query character vector. Allows users to add their own ACLED API queries to the
 #' GET call. Vector elements are assumed to be individual queries, and are automatically separated by an & sign.
 #' See [ACLED's API Guide](https://apidocs.acleddata.com/) for information on making your custom query.